[add] history for cut and paste

+++ src/services/constructor/HistoryService.ts
@@ -124,11 +124,7 @@ export default class HistoryService {
    * @param oldVal Old value
    * @param newVal New value
    */
-  registerElementMove(
-    id: string,
-    oldVal: { top: number; left: number },
-    newVal: { top: number; left: number }
-  ) {
+  registerElementMove(id: string, oldVal: { top: number; left: number }, newVal: { top: number; left: number }) {
     let isChanged = false
     for (const key in oldVal) {
       if (oldVal[key] != newVal[key]) isChanged = true
@@ -136,12 +132,7 @@ export default class HistoryService {
     if (isChanged) this.registerAction('moveElement', { id }, oldVal, newVal)
   }
 
-  registerPropertyChange(
-    id: string,
-    propertyName: string,
-    oldVal: any,
-    newVal: any
-  ) {
+  registerPropertyChange(id: string, propertyName: string, oldVal: any, newVal: any) {
     const old: any = {}
     old[propertyName] = oldVal
 
@@ -152,26 +143,11 @@ export default class HistoryService {
   }
 
   registerTextChange(id: string, oldVal: string, newVal: string) {
-    this.registerAction(
-      'textChange',
-      { id },
-      { content: oldVal },
-      { content: newVal }
-    )
+    this.registerAction('textChange', { id }, { content: oldVal }, { content: newVal })
   }
 
-  registerColorCorrection(
-    id: string,
-    parameter: string,
-    oldVal: string,
-    newVal: string
-  ) {
-    this.registerAction(
-      'colorCorrectionChange',
-      { id },
-      { [parameter]: oldVal },
-      { [parameter]: newVal }
-    )
+  registerColorCorrection(id: string, parameter: string, oldVal: string, newVal: string) {
+    this.registerAction('colorCorrectionChange', { id }, { [parameter]: oldVal }, { [parameter]: newVal })
   }
 
   registerSlideDelete(index: number, slideData: Map<string, ISlideObject>) {
@@ -180,12 +156,7 @@ export default class HistoryService {
   }
 
   registerSlideCreate() {
-    this.registerAction(
-      'createSlide',
-      { index: presentation.slides.length },
-      null,
-      null
-    )
+    this.registerAction('createSlide', { index: presentation.slides.length - 1 }, null, null)
   }
 
   registerElementDelete(elements: ISlideObject[], slideIndex: number) {
@@ -193,29 +164,30 @@ export default class HistoryService {
     for (const item of elements) {
       ids.push(item.id)
     }
-    this.registerAction(
-      'deleteElement',
-      { id: ids.join(';'), index: slideIndex },
-      elements,
-      null
-    )
+    this.registerAction('deleteElement', { id: ids.join(';'), index: slideIndex }, elements, null)
+  }
+
+  registerCut(elements: ISlideObject[], slideIndex: number) {
+    const ids: string[] = []
+    for (const item of elements) {
+      ids.push(item.id)
+    }
+    this.registerAction('cut', { id: ids.join(';'), index: slideIndex }, elements, null)
+  }
+
+  registerPaste(elements: ISlideObject[], slideIndex: number) {
+    const ids: string[] = []
+    for (const item of elements) {
+      ids.push(item.id)
+    }
+    this.registerAction('paste', { id: ids.join(';'), index: slideIndex }, elements, null)
   }
 
   registerElementCreate(element: ISlideObject, slideIndex: number) {
-    this.registerAction(
-      'createElement',
-      { id: element.id, index: slideIndex },
-      element,
-      null
-    )
+    this.registerAction('createElement', { id: element.id, index: slideIndex }, element, null)
   }
 
-  async registerAction(
-    actionType: string,
-    waybackArgs: IWaybackProperties,
-    oldValue: any,
-    newValue: any
-  ) {
+  async registerAction(actionType: string, waybackArgs: IWaybackProperties, oldValue: any, newValue: any) {
     console.log('registering action')
     const newRecord: IHistoryRecord = {
       actionType,
@@ -244,10 +216,14 @@ export default class HistoryService {
         newRecord.waybackFunction = 'createSlideWayback'
         break
       case 'deleteElement':
+      case 'cut':
         newRecord.waybackFunction = 'deleteElementWayback'
         break
       case 'createElement':
         newRecord.waybackFunction = 'createElementWayback'
+        break
+      case 'paste':
+        newRecord.waybackFunction = 'createMultipleElementsWayback'
         break
     }
 
@@ -274,10 +250,7 @@ export default class HistoryService {
   changeElementWayback(record: IHistoryRecord, reverse?: boolean) {
     //args: any, oldValue: object, newValue: object
     const id = record.waybackArguments?.id ?? 'null'
-    slideObjectService.changeObjectProperties(
-      id,
-      reverse ? record.newValue : record.oldValue
-    )
+    slideObjectService.changeObjectProperties(id, reverse ? record.newValue : record.oldValue)
   }
 
   deleteSlideWayback(record: IHistoryRecord, reverse?: boolean) {
@@ -286,12 +259,10 @@ export default class HistoryService {
       const oldSlideArray = presentation.slides
       for (let i = 0; i < oldSlideArray.length; i++) {
         const slide = oldSlideArray[i]
-        if (i == record.waybackArguments.index)
-          newSlideArray.push(record.oldValue)
+        if (i == record.waybackArguments.index) newSlideArray.push(record.oldValue)
         newSlideArray.push(slide)
       }
-      if (record.waybackArguments.index == oldSlideArray.length)
-        newSlideArray.push(record.oldValue)
+      if (record.waybackArguments.index == oldSlideArray.length) newSlideArray.push(record.oldValue)
       presentation.slides = newSlideArray
     } else {
       slideService.deleteSlide(record.waybackArguments.index ?? 0)
@@ -300,8 +271,7 @@ export default class HistoryService {
 
   createSlideWayback(record: IHistoryRecord, reverse?: boolean) {
     if (!reverse) {
-      if (record.waybackArguments.index)
-        slideService.deleteSlide(record.waybackArguments.index)
+      if (typeof record.waybackArguments.index == 'number') slideService.deleteSlide(record.waybackArguments.index)
     } else {
       slideService.createSlide()
     }
@@ -329,6 +299,21 @@ export default class HistoryService {
       slide.delete(record.waybackArguments.id ?? '')
     } else {
       slide.set(record.oldValue.id, record.oldValue)
+    }
+  }
+
+  createMultipleElementsWayback(record: IHistoryRecord, reverse?: boolean) {
+    const slide = presentation.slides[record.waybackArguments.index ?? 0]
+    if (!reverse) {
+      for (const item of record.oldValue) {
+        try {
+          slide.delete(item.id)
+        } catch {}
+      }
+    } else {
+      for (const item of record.oldValue) {
+        slide.set(item.id, item)
+      }
     }
   }
 }
